+++ src/yafowil/common.py
@@ -5,15 +5,12 @@ from yafowil.base import factory
 from yafowil.base import fetch_value
 from yafowil.compat import ITER_TYPES
 from yafowil.compat import STR_TYPE
-from yafowil.datatypes import convert_value_to_datatype
-from yafowil.datatypes import convert_values_to_datatype
 # alias generic_datatype_extractor and generic_emptyvalue_extractor imports
 # for now to make deprecated import warnings work
 from yafowil.datatypes import generic_datatype_extractor as generic_datatype_extractor_
 from yafowil.datatypes import generic_emptyvalue_extractor as generic_emptyvalue_extractor_
 from yafowil.datatypes import lookup_datatype_converter
 from yafowil.tsf import _
-from yafowil.utils import EMPTY_VALUE
 from yafowil.utils import as_data_attrs
 from yafowil.utils import attr_value
 from yafowil.utils import css_managed_props
@@ -22,7 +19,6 @@ from yafowil.utils import cssid
 from yafowil.utils import managedprops
 from yafowil.utils import vocabulary
 from zope.deferredimport import deprecated
-import re
 
 
 ###############################################################################
@@ -107,6 +103,28 @@ deprecated(
 deprecated(
     '``number_extractor`` has been moved to ``yafowil.number``.',
     number_extractor='yafowil.number:number_extractor'
+)
+
+# password
+deprecated(
+    '``minlength_extractor`` has been moved to ``yafowil.password``.',
+    minlength_extractor='yafowil.password:minlength_extractor'
+)
+deprecated(
+    '``ascii_extractor`` has been moved to ``yafowil.password``.',
+    ascii_extractor='yafowil.password:ascii_extractor'
+)
+deprecated(
+    '``password_extractor`` has been moved to ``yafowil.password``.',
+    password_extractor='yafowil.password:password_extractor'
+)
+deprecated(
+    '``password_edit_renderer`` has been moved to ``yafowil.password``.',
+    password_edit_renderer='yafowil.password:password_edit_renderer'
+)
+deprecated(
+    '``password_display_renderer`` has been moved to ``yafowil.password``.',
+    password_display_renderer='yafowil.password:password_display_renderer'
 )
 
 # select
@@ -790,210 +808,6 @@ Flag textarea is readonly.
 """
 
 factory.defaults['lines.persist'] = True
-
-
-###############################################################################
-# password
-###############################################################################
-
-@managedprops('minlength')
-def minlength_extractor(widget, data):
-    """Validate minlength of a string input.
-
-    Only perform if ``minlength`` property is set.
-
-    Properties:
-
-    ``minlength``
-        Minimum length of string as int.
-    """
-    val = data.extracted
-    if val is UNSET:
-        return val
-    minlength = attr_value('minlength', widget, data, -1)
-    if minlength != -1:
-        if len(val) < minlength:
-            message = _('minlength_extraction_error',
-                        default=u'Input must have at least ${len} characters.',
-                        mapping={'len': minlength})
-            raise ExtractionError(message)
-    return val
-
-
-@managedprops('ascii')
-def ascii_extractor(widget, data):
-    """Validate if a string is ASCII encoding.
-
-    Only perform if ``ascii`` property evaludates to True.
-
-    Properties:
-
-    ``ascii``
-        Flag  ascii check should perform.
-    """
-    val = data.extracted
-    if val is UNSET:
-        return val
-    if not attr_value('ascii', widget, data, False):
-        return val
-    try:
-        val.encode('ascii')
-    except UnicodeEncodeError:
-        message = _(u'ascii_extractor_error',
-                    default=u'Input contains illegal characters.')
-        raise ExtractionError(message)
-    return val
-
-
-LOWER_CASE_RE = r'(?=.*[a-z])'
-UPPER_CASE_RE = r'(?=.*[A-Z])'
-DIGIT_RE = r'(?=.*[\d])'
-SPECIAL_CHAR_RE = r'(?=.*[\W])'
-RE_PASSWORD_ALL = [
-    LOWER_CASE_RE,
-    UPPER_CASE_RE,
-    DIGIT_RE,
-    SPECIAL_CHAR_RE]
-PASSWORD_NOCHANGE_VALUE = '_NOCHANGE_'
-
-
-@managedprops('strength', 'weak_password_message')
-def password_extractor(widget, data):
-    """Extract and validate password input.
-
-    If extracted password is unchanged, return ``UNSET``. Consider this when
-    reading from password widgets!
-
-    This extractor provides a strength check. It only performs if ``strenght``
-    property is set. Strength check is done by four rules:
-        - input contains lowercase character
-        - input contains uppercase character
-        - input contains digit
-        - input contains special character.
-
-    Properties:
-
-    ``strength``
-        Integer value <= 4. Define how many rules must apply to consider a
-        password valid.
-    """
-    val = data.extracted
-    if val == PASSWORD_NOCHANGE_VALUE:
-        return UNSET
-    if val is UNSET:
-        return val
-    required_strength = attr_value('strength', widget, data, 0)
-    if required_strength <= 0:
-        return val
-    if required_strength > len(RE_PASSWORD_ALL):
-        required_strength = len(RE_PASSWORD_ALL)
-    strength = 0
-    for reg_exp in RE_PASSWORD_ALL:
-        if re.match(reg_exp, val):
-            strength += 1
-    if strength < required_strength:
-        error = attr_value('weak_password_message', widget, data)
-        raise ExtractionError(error)
-    return val
-
-
-def _pwd_value(widget, data):
-    if data.extracted is not UNSET:
-        return data.extracted
-    if data.value is not UNSET \
-       and data.value is not None:
-        return PASSWORD_NOCHANGE_VALUE
-    return attr_value('default', widget, data)
-
-
-@managedprops(
-    'data',
-    'title',
-    'size',
-    'disabled',
-    'autofocus',
-    'placeholder',
-    'autocomplete',
-    *css_managed_props)
-def password_edit_renderer(widget, data):
-    """Render password widget.
-    """
-    tag = data.tag
-    input_attrs = input_attributes_common(widget, data)
-    input_attrs['type'] = 'password'
-    input_attrs['value'] = _pwd_value(widget, data)
-    return tag('input', **input_attrs)
-
-
-@managedprops('displayplaceholder')
-def password_display_renderer(widget, data):
-    value = _pwd_value(widget, data)
-    if value == PASSWORD_NOCHANGE_VALUE:
-        return attr_value('displayplaceholder', widget, data)
-    return u''
-
-
-factory.register(
-    'password',
-    extractors=[
-        generic_extractor,
-        generic_required_extractor,
-        generic_emptyvalue_extractor_,
-        minlength_extractor,
-        ascii_extractor,
-        password_extractor
-    ],
-    edit_renderers=[password_edit_renderer],
-    display_renderers=[password_display_renderer]
-)
-
-factory.doc['blueprint']['password'] = """\
-Password blueprint.
-
-The password is never rendered to markup, instead
-``yafowil.common.PASSWORD_NOCHANGE_VALUE`` is set as ``value`` property on
-dom element. See ``yafowil.common.password_extractor`` for details on
-password extraction.
-"""
-
-factory.defaults['password.required_class'] = 'required'
-
-factory.defaults['password.default'] = ''
-
-factory.defaults['password.class'] = 'password'
-
-factory.defaults['password.minlength'] = -1
-factory.doc['props']['password.size'] = """\
-Maximum length of password.
-"""
-
-factory.doc['props']['password.minlength'] = """\
-Minimum length of password.
-"""
-
-factory.defaults['password.ascii'] = False
-factory.doc['props']['password.ascii'] = """\
-Flag ascii check should performed.
-"""
-
-factory.defaults['password.strength'] = -1
-factory.doc['props']['password.strength'] = """\
-Integer value <= 4. Define how many rules must apply to consider a password
-valid.
-"""
-
-factory.defaults['weak_password_message'] = _('weak_password_message',
-                                              default=u'Password too weak')
-factory.doc['props']['password.weak_password_message'] = """\
-Message shown if password is not strong enough.
-"""
-
-factory.defaults['password.displayplaceholder'] = u'*' * 8
-factory.doc['props']['password.displayplaceholder'] = """\
-Placeholder shown in display mode if password was set.
-"""
-
-factory.defaults['password.persist'] = True
 
 
 ###############################################################################
